requirements.txt
beautifulsoup4==4.11.1
boto3==1.26.89
botocore==1.29.89
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==39.0.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.24.2
protobuf==3.20.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.0
s3transfer==0.6.0
scipy==1.10.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.5.0
urllib3==1.26.9
Werkzeug==2.1.2
